numpy v2: refactor previous commit
should avoid passing around np.int64 items when it's supposed to be 32
bit items. the int64 is coming from doing computations on a mix of int32
and uint32 and numpy upcasting to int64 to avoid potential "precision loss"

+++ src/simplemseed/steim2.py
@@ -509,4 +509,4 @@ def steimPackWord(diff: list[int], nbits: int, ndiff: int, bitmask: int, submask
         i += 1
     if submask != 0:
         val |= submask << 30
-    return val
+    return val.astype(numpy.int32)
